+++ src/main/java/MasterThesis/arc/ArcFactory.java
@@ -17,7 +17,7 @@ public class ArcFactory {
         entity.setStartNodeId(Long.valueOf(entityArray[1])); //Pocz //TODO Powinno wskazywać na obiekt Node
         entity.setEndNodeId(Long.valueOf(entityArray[2])); //Konc   //TODO Powinno wskazywać na obiekt Node
         entity.setType(ArcType.valueOf(entityArray[3]));     //Typ
-        entity.setPosition(Integer.valueOf(entityArray[4]));      //Pozycja
+        entity.setPosition(Long.valueOf(entityArray[4]));      //Pozycja
         entity.setArcLength(Double.valueOf(entityArray[5]));    //Dlugosc
         entity.setLines(Integer.valueOf(entityArray[6]));        //Tory
         entity.setState(Integer.valueOf(entityArray[7]));        //Stan
